Move transaction class to Clauses folder
