+++ src/QueryBuilder/QueryBuilder.php
@@ -9,7 +9,7 @@ use Saraf\QB\QueryBuilder\Clauses\Insert;
 use Saraf\QB\QueryBuilder\Clauses\InsertUpdate;
 use Saraf\QB\QueryBuilder\Clauses\MultiInsertUpdate;
 use Saraf\QB\QueryBuilder\Clauses\Select;
-use Saraf\QB\QueryBuilder\Clauses\Transaction\Transaction;
+use Saraf\QB\QueryBuilder\Clauses\Transaction;
 use Saraf\QB\QueryBuilder\Clauses\Update;
 use Saraf\QB\QueryBuilder\Core\DBFactory;
